+++ GrowDiary.Web/Program.cs
@@ -84,6 +84,7 @@ builder.Services.AddScoped<AlertEvaluationService>();
 builder.Services.AddSingleton<NotificationSettingsRepository>();
 builder.Services.AddSingleton<AiSettingsRepository>();
 builder.Services.AddSingleton<AppSettingsRepository>();
+builder.Services.AddSingleton<SolutionStabilityAnalyzer>();
 builder.Services.AddScoped<TrendWatchRunner>();
 builder.Services.AddSingleton<AiClient>();
 builder.Services.AddSingleton<AiContextBuilder>();